+++ glamod_marine_processing/obs_suite/scripts/level1b.py
@@ -73,6 +73,16 @@ from cdm_reader_mapper.operations import replace
 reload(logging)  # This is to override potential previous config of logging
 
 
+def drop_qualities(df, drop_dict):
+    """Drop rows with bad quality flags."""
+    for column, values in drop_dict.items():
+        if not isinstance(values, list):
+            values = [values]
+        df = df[~df[column].isin(values)]
+
+    return df
+
+
 # MAIN ------------------------------------------------------------------------
 # Process input and set up some things ----------------------------------------
 logging.basicConfig(
@@ -88,7 +98,13 @@ else:
     logging.error("Need arguments to run!")
     sys.exit(1)
 
-process_options = ["correction_version", "corrections", "histories"]
+process_options = [
+    "correction_version",
+    "corrections",
+    "histories",
+    "duplicates",
+    "drop_qualities",
+]
 params = script_setup(process_options, args, "level1b", "level1a")
 
 cor_ext = ".txt.gz"
@@ -118,6 +134,7 @@ for table in cdm.properties.cdm_tables:
     table_df = cdm.read_tables(
         params.prev_level_path, params.prev_fileID, cdm_subset=[table]
     )
+
     if len(table_df) == 0:
         logging.warning(f"Empty or non-existing table {table}")
         correction_dict[table]["read"] = 0
@@ -203,7 +220,7 @@ for table in cdm.properties.cdm_tables:
         if table == "header":
             table_df["history"].loc[replaced] = (
                 table_df["history"].loc[replaced]
-                + f";{history_tstmp}. {params.histories.get(correction)}"
+                + f"; {history_tstmp}. {params.histories.get(correction)}"
             )
 
         table_df.drop(element + ".former", axis=1)
@@ -212,6 +229,12 @@ for table in cdm.properties.cdm_tables:
     # Track duplicate status
     if table == "header":
         correction_dict["duplicates"] = {}
+        if params.correction_version == "null":
+            if params.drop_qualities:
+                table_df = drop_qualities(table_df, params.drop_qualities)
+            DupDetect = cdm.duplicate_check(table_df, **params.duplicates)
+            DupDetect.flag_duplicates()
+            table_df = DupDetect.result
         contains_info = table_df["duplicate_status"] != dupNotEval
         logging.info("Logging duplicate status info")
         if len(np.where(contains_info)[0]) > 0: