+++ glamod_marine_processing/obs_suite/scripts/level1c.py
@@ -154,7 +154,10 @@ def read_table_files(table):
     #    prev_level_path, fileID, cdm_subset=[table], na_values="null"
     # )
     table_df = cdm.read_tables(
-        params.prev_level_path, cdm_subset=[table], na_values="null"
+        params.prev_level_path,
+        params.prev_fileID,
+        cdm_subset=[table],
+        na_values="null",
     )
     try:
         len(table_df)