+++ tests/_testing_suite.py
@@ -9,12 +9,9 @@ from cdm_reader_mapper.cdm_mapper import read_tables
 from cdm_reader_mapper.common.getting_files import load_file
 
 add_data = {
-    "level1a": None,
-    "level1b": _load_data.load_NOC_corrections,
     "level1c": _load_data.load_NOC_ANC_INFO,
     "level1d": _load_data.load_Pub47,
     "level1e": _load_data.load_metoffice_qc,
-    "level2": None,
 }
 
 
@@ -28,13 +25,15 @@ def _obs_testing(dataset, level, capsys):
         if level in _settings.manipulation.keys():
             for index, values in _settings.manipulation[level].items():
                 expected[index] = values
+        if not hasattr(_settings, "drops"):
+            return expected
         if level in _settings.drops.keys():
             expected = expected.drop(_settings.drops[level]).reset_index(drop=True)
         return expected
 
     _settings = get_settings(dataset)
     tables = _settings.which_tables[level]
-    if add_data[level] is not None:
+    if add_data.get(level) is not None:
         add_data[level](
             cache_dir=f"./T{level}/release_7.0",
         )